Move proptype generation
