+++ .watchmanconfig
@@ -1,0 +1,3 @@
+{
+  "root_files": [".watchmanconfig"]
+}
